Restructure example for clarity.
