+++ examples/features/authentication/pushed-authorization-request/composer.json
@@ -2,7 +2,7 @@
   "name": "auth0-samples/pushed-authorization-request",
   "description": "Example of using Pushed Authorization Requests with Auth0 PHP",
   "require": {
-    "php": "^8.0",
+    "php": "^8.1",
     "auth0/auth0-php": "^8.6",
     "nyholm/psr7": "^1.5",
     "symfony/http-client": "^6.2"
@@ -19,7 +19,7 @@
   },
   "scripts": {
     "serve": [
-      "php -S 127.0.0.1:3000 src/index.php"
+      "php -S 127.0.0.1:3000 index.php"
     ],
     "pre-install-cmd": [
       "@php -r \"file_exists('.env') || copy('.env.example', '.env');\""
