+++ examples/features/authentication/pushed-authorization-request/index.php
@@ -2,36 +2,61 @@
 
 declare(strict_types=1);
 
-use Auth0\SDK\Auth0;
-use Auth0\SDK\Configuration\SdkConfiguration;
+/**
+ * This example demonstrates how to use the Auth0 SDK for PHP to perform a Pushed Authorization Request (PAR).
+ *
+ * You should invoke this app using routed URLs like http://localhost:3000 or http://localhost:3000/login.
+ * Using the PHP built-in web server, you can start this app with the following command: php -S 127.0.0.1:3000 index.php
+ */
+
 use Auth0\SDK\Contract\Auth0Interface;
 
-require __DIR__ . '/vendor/autoload.php';
+// Review the common.php file for SDK setup and configuration details.
+require __DIR__ . '/../common.php';
 
+// Update the SDK configuration (created in ../common.php) to use Pushed Authorization Requests.
+$sdk->configuration()->setPushedAuthorizationRequest(true);
+
+// To avoid confusion between 127.0.0.1 and localhost (which are not interchangeable as far as cookies are concerned) we'll only allow this example to run on localhost.
+// Comment out this call if you want to run this example on a different domain.
 requireLocalhost();
 
-$sdk = setupSdk();
-
+// Setup routing.
 switch (getRoute()[0] ?? '') {
     case 'login':
-        routeLogin($sdk);
+        login($sdk);
         break;
 
     case 'logout':
-        routeLogout($sdk);
+        logout($sdk);
         break;
 
     case 'callback':
-        routeCallback($sdk);
+        callback($sdk);
         break;
 
     default:
-        routeDefault($sdk);
+        home($sdk);
         break;
 }
 
-function routeDefault(Auth0Interface $sdk): void {
+// Index/homepage route. Just shows the user's profile if logged in, or a login link if not.
+function home(Auth0Interface $sdk): void {
+    // Check if the user is authenticated.
     if ($sdk->isAuthenticated()) {
+        // Check if the user's access token has expired.
+        if (true === $sdk->getCredentials()?->accessTokenExpired) {
+            try {
+                $sdk->renew(); // Attempt to renew the session using an available refresh token.
+            } catch (Throwable $exception) {
+                $sdk->clear(); // If the session cannot be renewed, clear it.
+            }
+
+            header("Refresh: 0");
+            echo("<meta http-equiv='refresh' content='0'>");
+            exit;
+        }
+
         echo '<p><pre>' . print_r($sdk->getUser(), true) . '</pre></p>';
         echo('<a href="/logout">Logout</a>');
     } else {
@@ -40,7 +65,8 @@ function routeDefault(Auth0Interface $sdk): void {
     }
 }
 
-function routeLogin(Auth0Interface $sdk): void {
+// Login route. Redirects to Auth0 for login.
+function login(Auth0Interface $sdk): void {
     $callback = getBaseUrl() . '/callback';
     $url = $sdk->login(redirectUrl: $callback);
 
@@ -48,62 +74,18 @@ function routeLogin(Auth0Interface $sdk): void {
     exit;
 }
 
-function routeLogout(Auth0Interface $sdk): void {
+// Logout route. Redirects to Auth0 for logout. Returns to the homepage after logout.
+function logout(Auth0Interface $sdk): void {
     header('Location: ' . $sdk->logout(returnUri: getBaseUrl()));
     exit;
 }
 
-function routeCallback(Auth0Interface $sdk): void {
+// Callback route. Handles the user's return trip from Auth0 after login.
+function callback(Auth0Interface $sdk): void {
     if ($sdk->getExchangeParameters()) {
         $sdk->exchange(redirectUri: getBaseUrl());
     }
 
     header('Location: /');
     exit;
-}
-
-function setupSdk(): Auth0Interface {
-    $env = parse_ini_file('.env', true, INI_SCANNER_TYPED);
-
-    $configuration = new SdkConfiguration(
-        domain: $env['DOMAIN'] ?? null,
-        customDomain: $env['CUSTOM_DOMAIN'] ?? null,
-        clientId: $env['CLIENT_ID'] ?? null,
-        clientSecret: $env['CLIENT_SECRET'] ?? null,
-        cookieSecret: $env['COOKIE_SECRET'] ?? null,
-        cookieExpires: 3600, // Session will expire in 1 hour
-        audience: $env['API_IDENTIFIER'] !== null && $env['API_IDENTIFIER'] !== '' ? [$env['API_IDENTIFIER']] : null,
-        pushedAuthorizationRequest: true,
-    );
-
-    return new Auth0($configuration);
-}
-
-function requireLocalhost() {
-    $currentUrl = getBaseUrl();
-
-    if (parse_url($currentUrl, PHP_URL_HOST) !== 'localhost') {
-        $redirectUrl = parse_url($currentUrl, PHP_URL_SCHEME);
-        $redirectUrl .= '://localhost';
-        $redirectUrl .= parse_url($currentUrl, PHP_URL_PORT) !== null ? ':' . parse_url($currentUrl, PHP_URL_PORT) : '';
-
-        header('Location: ' . $redirectUrl);
-        exit;
-    }
-}
-
-function getBaseUrl(): string {
-    $currentUrl = isset($_SERVER['HTTPS']) && $_SERVER['HTTPS'] === 'on' ? 'https' : 'http';
-    $currentUrl .= '://' . $_SERVER['HTTP_HOST'] . $_SERVER['REQUEST_URI'];
-
-    $baseUrl = parse_url($currentUrl, PHP_URL_SCHEME);
-    $baseUrl .= '://' . parse_url($currentUrl, PHP_URL_HOST);
-    $baseUrl .= parse_url($currentUrl, PHP_URL_PORT) !== null ? ':' . parse_url($currentUrl, PHP_URL_PORT) : '';
-
-    return $baseUrl;
-}
-
-function getRoute(): array {
-    $currentUrl = 'http://' . $_SERVER['HTTP_HOST'] . $_SERVER['REQUEST_URI'];
-    return array_values(array_filter(explode('/', parse_url($currentUrl, PHP_URL_PATH) ?? '')));
 }
